style: entity 수정
- 필드 타입 변경
- HostDailySalesStats 생성자 빌더 수정

+++ src/main/java/com/refactoringhabit/stats/domain/entity/HostTotalSalesStats.java
@@ -34,23 +34,23 @@ public class HostTotalSalesStats extends BaseTimeEntity {
 
     @Setter
     @Column(name = "sales_volume")
-    private int salesVolume;
+    private Long salesVolume;
 
     @Setter
     @Column(name = "sales_amount")
-    private int salesAmount;
+    private Long salesAmount;
 
     @Setter
     @Column(name = "refund_count")
-    private int refundCount;
+    private Long refundCount;
 
     @Setter
     @Column(name = "review_count")
-    private int reviewCount;
+    private Long reviewCount;
 
     @Setter
     @Column(name = "review_average")
-    private int reviewAverage;
+    private Long reviewAverage;
 
     @Builder
     public HostTotalSalesStats(Long hostId, String altId) {
